+++ src/views/Messages.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
 export default function Messages() {
-  return <p>Messaes</p>;
+  return <p>Messages</p>;
 }
